refactor(client): extract useFileManagement from P2PTransfer
First step of decomposing the 1.5k-line P2PTransfer.tsx into focused hooks,
mirroring the cli/engine/ split. Move the sender's file-selection concern
(files, isDragging, totalBytes, and the pick/drag/drop/delete handlers) into
client/hooks/useFileManagement.ts.

Behavior-preserving: the component now consumes the hook. currentFileIndex
stays in the component because it tracks transfer progress, not selection.
The Playwright e2e (which drives file selection via setInputFiles) is the
guard that behavior is unchanged.
